libsql: Add initial feature flag support
This PR adds working feature flag support for libsql. There are now 4
new features `core`, `replication`, `hrana`, `serde` (these names are subject
to change). By default, all features are enabled and explicitly `core`,
`replication` and `hrana` are and `serde` is a sub-dependent of these.

The idea behind these features is to enable users to use the hrana based
connection without needing to link libsql-sys, etc.

In addition, there is a new check added to `make testlibsql` that runs
`cargo hack` to verify that feature flags are setup correctly in CI and
locally.
