+++ crates/bindings/wasm/Cargo.toml
@@ -8,5 +8,5 @@ crate-type = ["cdylib"]
 
 [dependencies]
 js-sys = "0.3.64"
-libsql = { path = "../../core", default_features = false }
+libsql = { path = "../../core", default_features = false, features = ["core"] }
 wasm-bindgen = "0.2"
