+++ crates/core/src/errors.rs
@@ -25,7 +25,7 @@ pub enum Error {
     #[error("Write delegation: `{0}`")]
     WriteDelegation(crate::BoxError), // Not in rusqlite
     #[error("bincode: `{0}`")]
-    Bincode(#[from] bincode::Error),
+    Bincode(crate::BoxError),
     #[error("invalid column index")]
     InvalidColumnIndex,
     #[error("invalid column type")]
@@ -68,4 +68,11 @@ pub fn error_from_code(code: i32) -> String {
 pub fn sqlite_errmsg_to_string(errmsg: *const std::ffi::c_char) -> String {
     let errmsg = unsafe { std::ffi::CStr::from_ptr(errmsg) }.to_bytes();
     String::from_utf8_lossy(errmsg).to_string()
+}
+
+#[cfg(feature = "replication")]
+impl From<bincode::Error> for Error {
+    fn from(e: bincode::Error) -> Self {
+        Error::Bincode(e.into())
+    }
 }
