+++ crates/core/src/hrana/mod.rs
@@ -3,8 +3,8 @@
 mod pipeline;
 mod proto;
 
+use crate::value::ValueType;
 use hyper::header::AUTHORIZATION;
-use libsql_sys::ValueType;
 use pipeline::{
     ClientMsg, Response, ServerMsg, StreamBatchReq, StreamExecuteReq, StreamRequest,
     StreamResponse, StreamResponseError, StreamResponseOk,
@@ -14,6 +14,7 @@ use proto::{Batch, BatchResult, Col, Stmt, StmtResult};
 use hyper::StatusCode;
 use hyper_rustls::{HttpsConnector, HttpsConnectorBuilder};
 
+use crate::database::ConnectorService;
 use crate::Error;
 use crate::{params::Params, Column};
 use std::collections::{HashMap, VecDeque};
@@ -21,7 +22,7 @@ use std::sync::atomic::{AtomicI64, AtomicU64, Ordering};
 use std::sync::{Arc, RwLock};
 
 use super::rows::{RowInner, RowsInner};
-use crate::connection::{Conn, ConnectorService};
+use crate::connection::Conn;
 use crate::transaction::Transaction;
 
 type Result<T> = std::result::Result<T, HranaError>;
