+++ crates/core/src/replication/connection.rs
@@ -2,7 +2,6 @@
 
 use std::sync::Arc;
 
-use libsql_sys::ValueType;
 use parking_lot::Mutex;
 
 use crate::replication::pb::{
@@ -11,12 +10,14 @@ use crate::replication::pb::{
 use crate::rows::{RowInner, RowsInner};
 use crate::statement::Stmt;
 use crate::transaction::Tx;
+use crate::value::ValueType;
 use crate::{
     params::Params, replication::Writer, Error, Result, Statement, Transaction, TransactionBehavior,
 };
 use crate::{Column, Row, Rows, Value};
 
-use crate::connection::{Conn, LibsqlConnection};
+use crate::connection::Conn;
+use crate::local::impls::LibsqlConnection;
 
 use super::parser::{self, StmtKind};
 use super::pb::{ExecuteResults, ResultRows};
@@ -575,7 +576,8 @@ impl RowsInner for RemoteRows {
         let col = self.0.column_descriptions.get(idx as usize).unwrap();
         col.decltype
             .as_deref()
-            .and_then(ValueType::from_str)
+            .and_then(libsql_sys::ValueType::from_str)
+            .map(ValueType::from)
             .ok_or(Error::InvalidColumnType)
     }
 }
@@ -607,7 +609,8 @@ impl RowInner for RemoteRow {
         let col = self.1.get(idx as usize).unwrap();
         col.decltype
             .as_deref()
-            .and_then(ValueType::from_str)
+            .and_then(libsql_sys::ValueType::from_str)
+            .map(ValueType::from)
             .ok_or(Error::InvalidColumnType)
     }
 }
